+++ src/modules/jinro/classes/Jinro.ts
@@ -1,5 +1,6 @@
 import dabyss = require("../../dabyss");
 import { DocumentClient } from "aws-sdk/clients/dynamodb";
+import { Player } from "../classes/JinroPlayer"
 
 
 const gameTable = process.env.gameTable;
@@ -22,32 +23,15 @@ export interface PositionNames {
 	citizen: string;
 }
 
-/**
- * (「・ω・)「 がおー
- *
- * @export
- * @class Jinro
- * @extends {Game}
- */
 export class Jinro extends dabyss.Game {
-	settingNames: string[];
-	defaultSettingStatus: boolean[];
+    settingNames: string[];
+    defaultSettingStatus: boolean[];
+    talkType: number;
+    positionNames: PositionNames;
+    positionNumbers: PositionNumbers;
+    players: Player[];
 
-	positionNames: PositionNames;
-
-	talkType: number;
-	isAliveStatus: boolean[];
-	positionNumbers: PositionNumbers;
-
-	/**
-	 * Jinroインスタンス作成
-	 *
-	 * @constructor
-	 * @extends Game
-	 * @param {string} groupId
-	 * @memberof Jinro
-	 */
-	constructor(groupId: string) {
+    constructor(groupId: string) {
 		super(groupId);
 		this.settingNames = ["type", "timer"];
 		this.defaultSettingStatus = [true, true];
@@ -70,17 +54,11 @@ export class Jinro extends dabyss.Game {
 			hunter: 0,
 			citizen: 0,
 		};
-	}
-
-	/**
-	 * 初期化
-	 *
-	 * @returns {Promise<void>}
-	 * @memberof Jinro
-	 */
-	async init(): Promise<void> {
-		try {
-			const data: DocumentClient.QueryOutput = await dabyss.dynamoQuery(
+    }
+    
+    async init(): Promise<void> {
+        try {
+            const data: DocumentClient.QueryOutput = await dabyss.dynamoQuery(
 				gameTable,
 				"group_id",
 				this.groupId,
@@ -97,7 +75,11 @@ export class Jinro extends dabyss.Game {
 							group_id: this.groupId,
 							game_id: this.gameId,
 						};
-						// this.userIds = game.user_ids as string[];
+						this.userIds = game.user_ids as string[];
+						// TODO: ここ合ってる？
+						for (let i = 0; i < this.userIds.length; i++) {
+							this.players[i] = new Player(this.userIds[i]);
+						}
 						this.day = game.day as number;
 						this.gameName = game.game_name as string;
 						this.gameStatus = game.game_status as string;
@@ -108,29 +90,22 @@ export class Jinro extends dabyss.Game {
 					}
 				}
 			}
-		} catch (err) {
-			console.error(err);
-			console.error("gameの初期化失敗");
-		}
-	}
 
-	/**
-	 * インスタンス作成
-	 *
-	 * @static
-	 * @param {string} groupId
-	 * @returns {Promise<Jinro>}
-	 * @memberof Jinro
-	 */
-	static async createInstance(groupId: string): Promise<Jinro> {
-		const jinro: Jinro = new Jinro(groupId);
-		await jinro.init();
-		return jinro;
-	}
+        } catch (err) {
+            console.error(err);
+            console.error("JinroGameの初期化失敗")
+        }
+    }
 
-	async update(): Promise<void> {
-		const jinro = {
-			group_id: this.groupId,
+    static async createInstance(groupId: string): Promise<JinroGame> {
+        const jinrogame: JinroGame = new JinroGame(groupId);
+        await jinrogame.init();
+        return jinrogame;
+    }
+
+    async update(): Promise<void> {
+        const jinrogame = {
+            group_id: this.groupId,
 			game_id: this.gameId,
 			user_ids: this.userIds,
 			day: this.day,
@@ -139,18 +114,15 @@ export class Jinro extends dabyss.Game {
 			setting_status: this.settingStatus,
 			timer: this.timer,
 			winner: this.winner,
-			positions: this.positions,
+            talk_type: this.talkType,
+            players: this.players,
+        };
+        await dabyss.dynamoUpdate(gameTable, jinrogame)
+    }
 
-			talk_type: this.talkType,
-			is_alive_status: this.isAliveStatus,
-
-		};
-		await dabyss.dynamoUpdate(gameTable, jinro);
-	}
-
-	async updatePositionNumbers(): Promise<void> {
-		const userNumber: number = await this.getUserNumber();
-		if (userNumber < 5) {
+    async updatePositionNumbers(): Promise<void> {
+        const userNumber: number = await this.getUserNumber();
+        if (userNumber < 5) {
 			this.positionNumbers.werewolf = 1;
 		} else if (userNumber == 5) {
 			this.positionNumbers.werewolf = 1;
@@ -190,11 +162,10 @@ export class Jinro extends dabyss.Game {
 				this.positionNumbers.forecaster +
 				this.positionNumbers.psychic +
 				this.positionNumbers.hunter);
+        await this.update();
+    }
 
-		await this.update();
-	}
-
-	async updatePositions() {
+    async updatePositions() {
 		await this.updatePositionNumbers();
 		const positions: string[] = [];
 
@@ -216,7 +187,6 @@ export class Jinro extends dabyss.Game {
 		for (let i = 0; i < this.positionNumbers.citizen; i++) {
 			positions.push("市民");
 		}
-
 		// ランダム並べ替え
 		for (let i = positions.length - 1; i > 0; i--) {
 			const j = Math.floor(Math.random() * (i + 1));
@@ -224,44 +194,141 @@ export class Jinro extends dabyss.Game {
 			positions[i] = positions[j];
 			positions[j] = tmp;
 		}
-
-		this.positions = positions;
-
+        
+		for (let i = 0; i < this.players.length; i++) {
+            this.players[i].position = positions[i];    
+        }
 		await this.update();
-	}
-
-	async getPosition(userIndex: number): Promise<string> {
-		const position = this.positions[userIndex];
-		return position;
-	}
-
-	async updateTalkType(type: number): Promise<void> {
+    }
+    
+    async updateTalkType(type: number): Promise<void> {
 		this.talkType = type;
 		await this.update();
-	}
-
-	async updateDefaultAliveStatus(): Promise<void> {
-		for (let i = 0; i < this.userIds.length; i++) {
-			this.isAliveStatus[i] = true;
-			console.log(this.isAliveStatus);
+    }
+    
+    async updateDefaultAliveStatus(): Promise<void> {
+		for (let i = 0; i < this.players.length; i++) {
+			this.players[i].isAlive = true;
 		}
 		await this.update();
 	}
 
-	async isAlive(index: number): Promise<boolean> {
-		return this.isAliveStatus[index];
-	}
-
-	async die(index: number): Promise<void> {
-		this.isAliveStatus[index] = false;
+	async updateDefaultReadyStatus(): Promise<void> {
+		for (let i = 0; i < this.players.length; i++) {
+			this.players[i].isReady = false;
+		}
 		await this.update();
 	}
 
-	async isWerewolf(index: number): Promise<boolean> {
-		const res: boolean = this.positions[index] == this.positionNames.werewolf;
-		return res;
+	async getAliveNumber(): Promise<number> {
+		let aliveNum = 0;
+		for (let i = 0; i < this.players.length; i++) {
+			if (this.players[i].isAlive) {
+				aliveNum++;
+			}
+		}
+		return aliveNum;
 	}
 
+	async getAliveWerewolfNumber(): Promise<number> {
+		let aliveNum = 0;
+		for (let i = 0; i < this.players.length; i++) {
+			const player: Player = this.players[i]
+			if (player.position == this.positionNames.werewolf) {
+				if (player.isAlive) {
+					aliveNum++;
+				}
+			}
+		}
+		return aliveNum;
+	}
+	
+	async getAliveNotWerewolfNumber(): Promise<number> {
+		const aliveNumber = await this.getAliveNumber();
+		const aliveWerewolfNumber = await this.getAliveWerewolfNumber();
+		return aliveNumber - aliveWerewolfNumber;
+	}
+
+	async isWerewolfWin(): Promise<boolean> {
+		const aliveNumber: number = await this.getAliveNumber();
+		const werewolfNumber: number = this.positionNumbers.werewolf;
+		const isWerewolfWin: boolean = aliveNumber - werewolfNumber <= werewolfNumber;
+		return isWerewolfWin;
+	}
+
+	async isCitizenWin(): Promise<boolean> {
+		const aliveWerewolfNumber = await this.getAliveWerewolfNumber();
+		return aliveWerewolfNumber == 0;
+	}
+
+	async getDeadPlayers(): Promise<Player[]> {
+		const deadPlayers: Player[] = [];
+		for (let i = 0; i < this.players.length; i++) {
+			const player = this.players[i];
+			if (!player.isAlive) {
+				deadPlayers.push(player);
+			}
+		}
+		return deadPlayers
+	}
+
+	async getAlivePlayersExceptOneself(oneself: Player): Promise<Player[]> {
+		const alivePlayers: Player[] = [];
+		for (let i = 0; i < this.players.length; i++) {
+			const player: Player = this.players[i]
+			if (player != oneself && player.isAlive) {
+				alivePlayers.push(player);
+			}
+		}
+		return alivePlayers;
+	}
+
+	async getAliveDisplayNamesExceptOneself(oneself: Player): Promise<string[]> {
+		const aliveDisplayNames: string[] = [];
+		const alivePlayers: Player[] = await this.getAlivePlayersExceptOneself(oneself);
+		for (let i = 0; i < alivePlayers.length; i++) {
+			aliveDisplayNames.push(alivePlayers[i].displayName);
+		}
+		return aliveDisplayNames;
+	}
+
+	async getAliveUserIndexesExceptOneself(oneself: Player): Promise<number[]> {
+		const aliveUserIndexes: number[] = [];
+		const alivePlayers: Player[] = await this.getAlivePlayersExceptOneself(oneself);
+		for (let i = 0; i < alivePlayers.length; i++) {
+			aliveUserIndexes.push(i);
+		}
+		return aliveUserIndexes;
+	}
+
+	async getDeadDisplayNamesExceptOneself(oneself: Player): Promise<string[]> {
+		const deadDisplayNames: string[] = [];
+		const deadPlayers: Player[] = await this.getDeadPlayers();
+		for (let i = 0; i < deadPlayers.length; i++) {
+			deadDisplayNames.push(deadPlayers[i].displayName);
+		}
+		return deadDisplayNames;
+	}
+
+	async getDeadIndexes(oneself: Player): Promise<number[]> {
+		const deadUserIndexes: number[] = [];
+		const deadPlayers: Player[] = await this.getDeadPlayers();
+		for (let i = 0; i < deadPlayers.length; i++) {
+			deadUserIndexes.push(i);
+		}
+		return deadUserIndexes;
+	}
+
+	async isAllMembersGetReady(): Promise<boolean> {
+		let isAllMembersGetReady: boolean = true;
+		for (let i = 0; i < this.players.length; i++) {
+			const player: Player = this.players[i];
+			if (!player.isReady) {
+				isAllMembersGetReady = false;
+			}
+		}
+		return isAllMembersGetReady;
+	}
 	async getWinnerIndexes(): Promise<number[]> {
 		const res: number[] = [];
 		for (let i = 0; i < this.positions.length; i++) {
@@ -282,105 +349,20 @@ export class Jinro extends dabyss.Game {
 		return res;
 	}
 
-	async getAliveNumber(): Promise<number> {
-		let aliveNum = 0;
-		for (const state of this.isAliveStatus) {
-			if (state) {
-				aliveNum++;
-			}
-		}
-		return aliveNum;
-	}
-
-	async getAliveWerewolfNumber(): Promise<number> {
-		let aliveNum = 0;
-		for (let i = 0; i < this.userIds.length; i++) {
-			if (this.positions[i] == this.positionNames.werewolf) {
-				if (this.isAliveStatus[i]) {
-					aliveNum++;
-				}
-			}
-		}
-		return aliveNum;
-	}
-
-	async getAliveNotWerewolfNumber(): Promise<number> {
-		const aliveNumber = await this.getAliveNumber();
-		const aliveWerewolfNumber = await this.getAliveWerewolfNumber();
-		return aliveNumber - aliveWerewolfNumber;
-	}
-
-	// TODO -> DONE 人狼側の勝利条件が満たされているか判定
-	async isWerewolfWin(): Promise<boolean> {
-		const aliveNumber: number = await this.getAliveNumber();
-		const werewolfNumber: number = this.positionNumbers.werewolf;
-		const isWerewolfWin: boolean = aliveNumber - werewolfNumber <= werewolfNumber;
-		return isWerewolfWin;
-	}
-
-	async isCitizenWin(): Promise<boolean> {
-		const aliveWerewolfNumber = await this.getAliveWerewolfNumber();
-		return aliveWerewolfNumber == 0;
-	}
-
-	async getDeadIndexes(): Promise<number[]> {
-		const deadIndexes: number[] = [];
-		for (let i = 0; i < this.userIds.length; i++) {
-			if (!this.isAliveStatus[i]) {
-				deadIndexes.push(i);
-			}
-		}
-		return deadIndexes;
-	}
-
 	async putAction() {
-		const userNumber: number = await this.getUserNumber();
-		const status: boolean[] = [];
-		for (let i = 0; i < userNumber; i++) {
+		for (let i = 0; i < this.players.length; i++) {
 			if (
-				this.positions[i] == this.positionNames.madman ||
-				this.positions[i] == this.positionNames.citizen ||
-				!this.isAliveStatus[i]
+				this.players[i].position == this.positionNames.madman ||
+				this.players[i].position == this.positionNames.citizen ||
+				!this.players[i].isAlive
 			) {
-				status[i] = true;
+				this.players[i].isReady = true
 			} else {
-				status[i] = false;
+				this.players[i].isReady = false
 			}
 		}
-		await Action.putAction(this.gameId, this.day, status);
+		await this.update();
 	}
 
-	async getAliveUserIndexesExceptOneself(index: number): Promise<number[]> {
-		const res: number[] = [];
-		for (let i = 0; i < this.userIds.length; i++) {
-			if (i != index && this.isAlive(i)) {
-				res.push(i);
-			}
-		}
-		return res;
-	}
-
-	async getAliveDisplayNamesExceptOneself(index: number): Promise<string[]> {
-		const res: string[] = [];
-		for (let i = 0; i < this.userIds.length; i++) {
-			if (i != index && (await this.isAlive(i))) {
-				const user: dabyss.User = new dabyss.User(this.userIds[i]);
-				const displayName: string = await user.getDisplayName();
-				res.push(displayName);
-			}
-		}
-		return res;
-	}
-
-	async getDeadDisplayNamesExceptOneself(index: number): Promise<string[]> {
-		const res: string[] = [];
-		for (let i = 0; i < this.userIds.length; i++) {
-			if (i != index && !(await this.isAlive(i))) {
-				const user: dabyss.User = new dabyss.User(this.userIds[i]);
-				const displayName: string = await user.getDisplayName();
-				res.push(displayName);
-			}
-		}
-		return res;
-	}
 }
+
